Fixed intermittent segfault in MembershipServiceTest.updateServerList_single.
It has only been reproduced on slower virtual machines. The problem is that
the CoordinatorService object and the CoordinatorUpdateManager object inside
are destructed before the CoordinatorServerList object. Thus, when the unit
test function reaches its end, the updater thread in CoordinatorServerList
may still access the update manager in CoordinatorServerList::pruneUpdates(),
which results in the segfault.

The solution is to embed the CoordinatorServerList object inside the
CoordinatorService object and declare it before the update manager.