Try to flush out the potential grant shortage problem/bug?
Now I tend to believe that it's not a problem that some client frequently
runs out of GRANTs. It's simply because there is an incast to the server
and that client's DATA packets are queued at the ToR. Since the server doesn't
get the DATA in time, it cannot send back the GRANT.

We can verify this by looking at numSchedMessagePacket and numGrantsGenerated.
They should be roughly equal, meaning that we are not granting less we should.