An overhaul of checkTimeouts() to eliminate spurious unexpected GRANTs.
In Homa, it's possible for a long message to be starved of grants for
quite some time if the receiver has higher priority messages to grant.
The old timeout checking logic doesn't take this into account and
delete the ServerRpc too eagerly resulting in unexpected GRANTs followed
by a server requested RESTART because the corresponding serverRpc cannot
be found. The correct behavior is to send PING occasionally when a
message is starved of grants.