Fixed a bug in the timing of the AllGather benchmark.
Previously only node 0 computed its elapsed time so the numbers were actually
node 0's completion time.

Also, redefine the stop time to be when the last merge completed (so we don't
count the time waiting for outgoing RPCs to finish). One might argue this is
cheating but I believe this is more appropriate for the micro-benchmark.