millisort: solved the mysterious ~20Gbps per-machine bi-directional BW bottleneck
The solution is to disable the per-context/endpoint PSM2 managed recv thread by
"export PSM2_RCVTHREAD=0". The exact reason why this limits the BW to 20Gbps is
unknown yet but I suspect incorrect thread affinity might be involved. Anyway,
this is not really interesting because we have our own dedicated comm. threads.

Another important change in this commit is allowing psm2 and/or libfabric to
use RAMCloud timetrace. See GNUmakefile for details.