ips4o config: reduced the default max_sequential_size to 1/4
Previously, max_sequentail_size was 7*512*4=14336 for 7 cores;
this cutoff value resulted in a weird hump in the "scaleDown" &
"scaleByTotalRecords" graphs in the paper. This is because when
records per server is smaller than that cutoff ips4o is essentially
only using 1 core. Our user-level threads are efficient enough
such that we can reduce the default min_parallel_blocks_per_thread
from 4 to 1.